requirements.txt
pyttsx3>=2.90
ultralytics
opencv-python
numpy
Pillow
SpeechRecognition
pocketsphinx

# Development/Testing
pytest